requirements.txt
anyio==4.9.0
certifi==2025.4.26
exceptiongroup==1.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.10
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.1
typing_extensions==4.13.2
yt-dlp==2025.3.31
